moves suite2p_wrapper to modules
