+++ src/ophys_etl/modules/suite2p_wrapper/schemas.py
@@ -1,18 +1,7 @@
-import datetime
-import os
-import pathlib
-import shutil
-import tempfile
-from typing import Dict, List, Optional
-
 import argschema
-import h5py
 import marshmallow as mm
-import suite2p
 
-
-class Suite2PWrapperException(Exception):
-    pass
+from ophys_etl.modules.suite2p_wrapper.utils import Suite2PWrapperException
 
 
 class Suite2PWrapperSchema(argschema.ArgSchema):
@@ -190,123 +179,3 @@ class Suite2PWrapperOutputSchema(argschema.schemas.DefaultSchema):
         values=argschema.fields.List(ExistingFile),
         required=True,
         description="retained output files from Suite2P")
-
-
-def copy_and_add_uid(
-        srcdir: pathlib.Path, dstdir: pathlib.Path,
-        basenames: List[str], uid: Optional[str] = None) -> Dict[str, str]:
-    """copy files matching basenames from a tree search of srcdir to
-    dstdir with an optional unique id inserted into the basename.
-
-    Parameters
-    ----------
-    srcdir : pathlib.Path
-       source directory
-    dstdir : pathlib.Path
-       destination directory
-    basenames : list
-        list of basenames to copy
-    uid : str
-        uid to insert into basename (example a timestamp string)
-
-    Returns
-    -------
-    copied_files : dict
-        keys are basenames and vaues are output paths as strings
-
-    """
-
-    copied_files: dict = {}
-
-    for basename in basenames:
-        result = list(srcdir.rglob(basename))
-        if len(result) == 0:
-            # case for copying *.tif when non-existent
-            continue
-        if len(result) != 1:
-            if "*" not in basename:
-                raise ValueError(f"{len(result)} matches found in {srcdir} "
-                                 f"for {basename}. Expected 1 match.")
-        copied_files[basename] = []
-        for iresult in result:
-            dstbasename = iresult.name
-            if uid is not None:
-                dstbasename = iresult.stem + f"_{uid}" + iresult.suffix
-
-            dstfile = dstdir / dstbasename
-
-            shutil.copyfile(iresult, dstfile)
-
-            copied_files[basename].append(str(dstfile))
-
-    return copied_files
-
-
-class Suite2PWrapper(argschema.ArgSchemaParser):
-    default_schema = Suite2PWrapperSchema
-    default_output_schema = Suite2PWrapperOutputSchema
-
-    def run(self):
-        self.logger.name = type(self).__name__
-        self.logger.setLevel(self.args.pop('log_level'))
-
-        # Should always exist as either a valid SHA or "unknown build"
-        # if running in docker container.
-        ophys_etl_commit_sha = os.environ.get("OPHYS_ETL_COMMIT_SHA",
-                                              "local build")
-        self.logger.info(f"OPHYS_ETL_COMMIT_SHA: {ophys_etl_commit_sha}")
-
-        # determine nbinned from bin_duration and movie_frame_rate
-        if 'nbinned' not in self.args:
-            with h5py.File(self.args['h5py'], 'r') as f:
-                nframes = f['data'].shape[0]
-            bin_size = (self.args['bin_duration']
-                        * self.args['movie_frame_rate'])
-
-            if bin_size > nframes:
-                raise Suite2PWrapperException(
-                    f"The desired frame bin duration "
-                    f"({self.args['bin_duration']} "
-                    f"seconds) and movie frame rate "
-                    f"({self.args['movie_frame_rate']} Hz) results in a bin "
-                    f"size ({bin_size} frames) larger than the number of "
-                    f"actual frames in the movie ({nframes})!")
-
-            self.args['nbinned'] = int(nframes / bin_size)
-            self.logger.info(f"Movie has {nframes} frames collected at "
-                             f"{self.args['movie_frame_rate']} Hz. To get a "
-                             f"bin duration of {self.args['bin_duration']} "
-                             f"seconds, setting nbinned to "
-                             f"{self.args['nbinned']}.")
-
-        # make a tempdir for Suite2P's output
-        with tempfile.TemporaryDirectory() as tdir:
-            self.args['save_path0'] = tdir
-            self.logger.info(f"Running Suite2P with output going to {tdir}")
-            suite2p.run_s2p(self.args)
-
-            self.logger.info(f"Suite2P complete. Copying output from {tdir} "
-                             f"to {self.args['output_dir']}")
-            # copy over specified retained files to the output dir
-            odir = pathlib.Path(self.args['output_dir'])
-            odir.mkdir(parents=True, exist_ok=True)
-            self.now = None
-            if self.args['timestamp']:
-                self.now = datetime.datetime.now().strftime('%Y%m%d%H%M%S%f')
-            output_files = copy_and_add_uid(
-                    pathlib.Path(tdir),
-                    odir,
-                    self.args['retain_files'],
-                    self.now)
-            for k, v in output_files.items():
-                self.logger.info(f"wrote {k} to {v}")
-
-        outdict = {
-                'output_files': output_files
-                }
-        self.output(outdict, indent=2)
-
-
-if __name__ == "__main__":  # pragma: no cover
-    s2pw = Suite2PWrapper()
-    s2pw.run()
