+++ tests/modules/suite2p_wrapper/resources/movie_100x100x100.h5
