+++ tests/pipelines/test_segment_postprocess_pipeline.py
@@ -9,13 +9,14 @@ import ophys_etl.transforms.postprocess_rois as post_rois
 
 import sys
 sys.modules['suite2p'] = Mock()
-import ophys_etl.transforms.suite2p_wrapper as s2pw  # noqa
+from ophys_etl.modules.suite2p_wrapper.schemas import \
+        Suite2PWrapperSchema, Suite2PWrapperOutputSchema  # noqa: E402
 import ophys_etl.pipelines.segment_postprocess_pipeline as sbpipe  # noqa
 
 
 class MockSuite2PWrapper(argschema.ArgSchemaParser):
-    default_schema = s2pw.Suite2PWrapperSchema
-    default_output_schema = s2pw.Suite2PWrapperOutputSchema
+    default_schema = Suite2PWrapperSchema
+    default_output_schema = Suite2PWrapperOutputSchema
 
     def run(self):
         stat = pathlib.Path(self.args['output_dir']) / "stat.npy"
